Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
qdrant-client>=1.10.0
pypdf>=5.0.0
youtube-transcript-api>=0.6.2
python-pptx>=1.0.2
